core-api: fix type inference of createRouteRef
